requirements.txt
streamlit
replicate
requests
Pillow
numpy
python-dotenv
